additions

+++ switch_os.sh
@@ -1,14 +1,19 @@
+set -x
 case $* in                                                                                                             
         ubuntu)                                                                                                     
 		# clean
 		find . -maxdepth 1 -type l -delete
 		ln -s  ubuntumate16.04/* .
-		rm sys
-		mkdir sys	
+		test -h sys && rm sys
+		test -d sys || mkdir sys	
                 ;;                                                              
 	slackware)
 		find . -maxdepth 1 -type l -delete
 		ln -s slackware-14.2/* .
+		;;
+	tinycore)
+		find . -maxdepth 1 -type l -delete
+		ln -s tinycore-9.0.3v7/* .	
 		;;
         '')                                                                                                            
 		echo "usage"
